Fixing installer

+++ installer/clear.sh
@@ -2,8 +2,8 @@
 
 echo "sh: cleaning..."
 
-find bin ! -name '.keep' -type f -exec rm -f {} +
-find dist ! -name '.keep' -type f -exec rm -f {} +
+[ -d "bin" ] && find bin ! -name '.keep' -type f -exec rm -f {} +
+[ -d "dist" ] && find dist ! -name '.keep' -type f -exec rm -f {} +
 find . ! -name '.keep' -wholename '*/tmp/*' -type f -exec rm -f {} +
 
 echo "sh: cleaning done"
